Fixed filetree format to zippable pack export
Now instead of being "what the pack looks like when you have it installed" it is "what the pack looks like on export", meaning that now you can make a zip file straight off the github and import that in, no manual effort needed. Unless you're adding more mods to the pack, in which case now there is more manual effort needed. Shhh.